responsividade pagina de erro
