+++ modulo3/projeto-labex/src/pages/AdminHomePage/AdminHomePage.jsx
@@ -1,67 +1,8 @@
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
-import styled from "styled-components";
-import { RequestTrips } from "../Hooks/RequestTrips";
+import { RequestTrips } from "../../Hooks/RequestTrips";
 import { useEffect, useState } from "react";
-
-const DivPainelAdm = styled.div`
-  display: flex;
-  width: 100vw;
-  height: 100vh;
-  flex-direction: column;
-  align-items: center;
-
-  h1{
-    color: white;
-    margin: 2%;
-  }
-`
-
-const ListTripsName = styled.div`
-  display: flex;
-  justify-content: space-between;
-  box-shadow: 1px 1px 5px gray;
-  padding: 1%;
-  font-weight: bold;
-  margin: 0.3% 0;
-  width: 30%;
-  background-color: whitesmoke;
-
-  :hover {
-    background-color: lightgray;
-  }
-
-  @media (max-width: 600px){
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 90%;
-  }
-`;
-
-const DivButtonsPanelAdm = styled.div`
-  display: flex;
-  width: 30%;
-  justify-content: space-between;
-
-  @media (max-width: 600px){
-    width: 100%;
-    justify-content: space-evenly;
-  }
-`
-
-const ButtonsAdm = styled.div`
-  display: flex;
-  justify-content: space-around;
-  width: 30%;
-  margin-bottom: 1%;
-
-  @media (max-width: 600px){
-    width: 100%;
-  }
-`
-
+import {DivPainelAdm, ListTripsName, DivButtonsPanelAdm, ButtonsAdm} from './styles'
 
 
 export default function AdminHomePage() {
@@ -89,7 +30,7 @@ export default function AdminHomePage() {
 
   const TripDetails = (id) => {
     localStorage.setItem('id', id)
-    navigate(`/admin/trips/${id}`);
+    navigate(`/admin/trips/details/${id}`);
     
   }
