+++ modulo3/projeto-labex/src/pages/ListTripsPage/styles.js
@@ -1,66 +1,22 @@
 import styled from "styled-components";
-import ImgSpaceCard from '../img/spacecard.png'
-import SpaceHome from '../img/spaceHome.jpeg'
-
-
-//HOME
-export const DivHome = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  height: 100vh;
-  background-image: url(${SpaceHome});
-  background-size: cover;
-
-    button {
-      margin: 10% 0;
-      width: 60%;
-      height: 50px;
-    }
-  
-
-  h1 {
-    color: cyan;
-    text-shadow: 1px 1px 5px orange;
-    font-size: 4em;
-  }
-
-  b{
-      color: orange;
-  }
-`;
-
-export const BottonImg = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 40%;
-
-  @media (max-width: 600px){
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-  }
-`
-
-export const ImgLogo = styled.img`
-  width: 50%;
-`
-
-//Lista de Viagens
+import ImgSpaceCard from '../../img/spacecard.png'
 
 export const DivListTrips = styled.div`
   display: flex;
+  width: 100%;
   flex-direction: column;
   align-items: center;
   margin: 2% 0;
+  margin-bottom: 10%;
 
   h2{
       text-align: center;
       margin: 2% 0;
       color: white;
+  }
+
+  @media (max-width: 600px){
+    margin-bottom: 100%;
   }
 `;
 
@@ -82,14 +38,13 @@ export const DivButtons = styled.div`
 export const GridCards = styled.div`
   display: grid;
   grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(4, 1fr);
   gap: 2%;
+  row-gap: 5%;
   width: 80%;
 
   @media (max-width: 600px){
     grid-template-columns: 1fr;
-    margin-bottom: 20%;
-    height: 100%;
+    margin: 5% 0;
   }
 `;
