+++ modulo3/projeto-labex/src/pages/TripDetailsPage/TripDetailsPage.jsx
@@ -1,51 +1,11 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import styled from "styled-components";
+import {DivDetailsPage, CardDetails, CardPending} from './styles.js'
 
-const DivDetailsPage = styled.div`
-  width: 100%;
-  height: 100vh;
-  color: white;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 2%;
-  text-align: center;
-  padding: 1%;
-
-  div {
-    margin: 5% auto;
-    width: 90%;
-  }
-
-  .divBottons {
-    display: flex;
-    border: none;
-    justify-content: space-around;
-    width: 80%;
-  }
-
-  @media (max-width: 600px){
-    grid-template-columns: 1fr;
-    margin-bottom: 50%;
-    height: 100%;
-  }
-`;
-
-const CardDetails = styled.div`
-  box-shadow: 1px 1px 10px white; 
-  padding: 2%;
-  text-align: justify;
-`;
-
-const CardPending = styled.div`
-  display: block;
-  box-shadow: 1px 1px 10px white;
-  padding: 2%;
-  text-align: justify;
-`;
 
 export default function TripDetailsPage() {
+  
   const navigate = useNavigate();
 
   const params = useParams();
@@ -62,8 +22,6 @@ export default function TripDetailsPage() {
 
     returnTripDetails();
   }, []);
-
-  // const id = localStorage.getItem("id");
 
   const returnTripDetails = () => {
     const aluno = "Roberto-Maia-Vaughan";
@@ -142,7 +100,7 @@ export default function TripDetailsPage() {
   };
 
   const mapApproved = approved.map((item) => {
-    return item.name;
+    return <li hey={item.id}>{item.name}</li>
   });
 
   const mapCandidates = candidates.map((item) => {
