+++ modulo3/projeto-labex/src/route/Routes.jsx
@@ -1,12 +1,12 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import HomePage from '../pages/HomePage'
-import ListTripsPage from "../pages/ListTripsPage";
-import ApplicationFormPage from '../pages/ApplicationFormPage'
-import LoginPage from '../pages/LoginPage'
-import AdminHomePage from '../pages/AdminHomePage'
-import CreateTripPage from '../pages/CreateTripPage'
-import TripDetailsPage from '../pages/TripDetailsPage'
-import ErrorPage from "../pages/ErrorPage";
+import HomePage from '../pages/HomePage/HomePage'
+import ListTripsPage from "../pages/ListTripsPage/ListTripsPage";
+import ApplicationFormPage from '../pages/ApplicationFormPage/ApplicationFormPage'
+import LoginPage from '../pages/LoginPage/LoginPage'
+import AdminHomePage from '../pages/AdminHomePage/AdminHomePage'
+import CreateTripPage from '../pages/CreateTripPage/CreateTripPage'
+import TripDetailsPage from '../pages/TripDetailsPage/TripDetailsPage'
+import ErrorPage from "../pages/ErrorPage/ErrorPage";
 
 export default function Router(){
     return(
@@ -18,7 +18,7 @@ export default function Router(){
               <Route exact path='/login' element={<LoginPage />}/>
               <Route exact path='/admin/trips/list' element={<AdminHomePage />}/>
               <Route exact path='/admin/trips/create' element={<CreateTripPage />}/>
-              <Route exact path='/admin/trips/:id' element={<TripDetailsPage />}/>
+              <Route exact path='/admin/trips/details/:id' element={<TripDetailsPage />}/>
               <Route path='*' element={<ErrorPage />}/>
           </Routes>
         </BrowserRouter>
